Over the past several days, we have updated our database models to reflect some requirements we couldn't predict when we initially designed the database. In this commit, we are adding a table that keeps track of User-Stage pairings, with due dates, dates completed, and related functionality. We now also have separate tables for Investigator-Study and Participant-Study relationships. For an idea of the models, consult portal/studies/models.py.
We have also learned how to override admin site save methods, making it easy to keep the database correctly updated as we make changes in the back end.

On the front end, we have the participant's single study view mostly finished, with navigation between stages and common behaviour. This template is undergoing testing as we add the Investigator's side.